Cast metric summary values to float to avoid numpy in checkpoints
PyTorch 2.6 defaults weights_only=True in torch.load, which rejects
numpy scalars. Converting to native float fixes future checkpoint loads.

+++ src/utils/metrics.py
@@ -93,7 +93,7 @@ class MetricTracker:
             self.values[k].append(v)
 
     def summary(self):
-        return {k: np.mean(v) for k, v in self.values.items()}
+        return {k: float(np.mean(v)) for k, v in self.values.items()}
 
     def __str__(self):
         s = self.summary()
